refactor(fsspec): remove upper bound on fsspec
We're using `pins` as a dependency in `ibis-framework` and ideally we
can remove the upper-pin on `fsspec` so that doesn't impact our users.
I think the usage of the moved `hash_name` function is the only breaking
change between the current upper pin and most recent release, so I've
added a small workaround to handle the change.

All of the local tests pass, I didn't yet set up sufficient credentials
to handle the various cloud tests.